#4 Vector list of Objects / Special Keys (Non-ACII keys)
Game Objects now using Abstract classes meaning that each element that it is referencing now has to have a Draw class inside of them. For example  GameObject*  (pointing to cube ) cube  has to have a draw() function available within its class or else it cannot be called or used and Has an error without opening the Engine.

Special Keys Now Additionally To have a functionality for the keyboard and see if they are called correctly by seeing them from the Console feed.

Vector list of Objects works by :
1. Including a Vector library
2. Making a gameobject vector list of objects
3. Adding a list of new objects.
4. Push back Objects from the list make into the stack
5. Call a draw function to Draw it to the Screen