#8 Mass  Update for Objects
Each Object is now represented with a mass, which can be applied to the object when assigning new objects within the main.cpp source file.

Mass is positioned first for every object and the logic of objects currently goes by as :
1. Mass
2. Position
3. Color
+ Of the object.

Objects Available :
1. Cone
2. Cube
3. Particle
4. Quad2D
5. Sphere

Possibilities of using these objects within the pool game is To apply Particles when a ball is collided

Cube or Quad could be useful for making a Que for pool or the tables platform within the game Scene.

Note :

- The Mass is defined within GameObject itself.