#15 An Update for Screen Information
Previously displayed text on the screen was char const*  which means there was either an option to display a static an unchanging char string or do a   "  const_cast   in order to convert or change it within the formula.

as well as it being necessary if I wish to use a pointer or object information it would have to be converted to a const char* or else it will not display and work.