#19 Rotation For RigidBody2D
Updates Variables now featuring specifically if it is a Linear function or Angular, for more clearability of the code and reading it.

- This change is updated for both Particle as well as Rigidbody2D

´--------------------

Rotation Logic :
1. Find the inertia which depends on the Object information and details, if it is a rectangle, circle triangle specifying its initial inertia.       (Requires Shape Information / Change in sin angle) for the calculation.

2. Calculate Forces -
Reset the Forces for rotation " AngularTorque "

Find Total Sum for Torque = Net Torque

From Net Torque divide it by inertia " Resistance of the object to find the acceleration of the object.

3. Find the velocity affecting the object
== /* Requires A proper Approach as defined here : https://cnx.org/contents/MymQBhVV@175.14:51fg7QFb@14/Angular-velocity I the booklet */ ==

5. Displacements should take into consideration of the Forces affecting from x axis and y axis.

Meaning that we need to calculate which force is more powerful to apply to the Orientation.

As an example :
velocity.x = 20f;
and
velocity.y = -18f;

Which means that velocity.x is stronger and the force applied to the orientation should be 2.0f  which has to be translated to radians or degrees to the orientation of the angle taking into consideration the previous already achieved angle.

Updated Functions of the Particle.cpp and header file

Which now Follow the same way as the Rigidbody except without any angular function.