#21 GameEngine  & RigidBody3D
Update !

Everything that was on main class has now been moved into GameEngine class instead, separating code for readability and in case any other type of engine is required or needed to be added.

Addition of
RigidBody3D

Now featuring a new class that currently has initials setup on it but no functional matters as of currently.