#27 Circle Inertia / Orientation
Now Featuring use of rigidbody2D moment of inertia of a circle

For this you will have to take a look the initialization in the Main.cpp

then Circle initialization and how it is all drawn, then go towards RigidBody2D where it chooses if the inertia calculation is rectangle one or circle version.

According to that the orientation calculation changes.

--
Made Some changes to Comments and grammar fixes.

--
Reset all Objects keyboard use and detail showing to false  To clear non necessary information displayed to screen as well as no keyboard use if that is not required..