#34 Sphere Sphere Collison Dedection
Initialized Sphere on Sphere Collision detection, right now the way it works currently there can only be Spheres on the Scene for it to work, and it even Checks if there is no Collision Added to the Object.

Consider adding, a new vector of objectsCollision GameObject   which would To the dedection between spheres only instead of other objects in the space as well which is a possibility as of currently.

Currently For loop inside of gameEngine.cpp only assumes that  Object 0 is player object and that the rest of the objects are that they collide with.