#35 Collision Check
Checking object 0 Against every other Object in the Scene, even objects that theoretically do not have a collision application on it.

Theoretically GameObjects Elements could all check if There is an attached Collider on the object or not, which would solve the problem of Checking collision with objects not necessary.